Serialization split by format versions
