+++ DBMerge.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Xml.Linq;
-
-namespace TranscriptionCore
+﻿namespace TranscriptionCore
 {
     public class DBMerge
     {
@@ -15,21 +9,6 @@ namespace TranscriptionCore
         {
             this.DBID = DBID;
             this.DBtype = type;
-        }
-
-        internal XElement Serialize()
-        {
-            string val;
-            if (this.DBtype == DBType.Api)
-                val = "api";
-            else if (DBtype == DBType.User)
-                val = "user";
-            else
-                val = "file";
-
-            return new XElement("m",
-                new XAttribute("dbid", DBID),
-                new XAttribute("dbtype",val));
         }
     }
 }
